requirements.txt
pdf2image>=1.17.0
Pillow>=10.3.0
ultralytics>=8.2.0
streamlit>=1.35.0
numpy>=1.26.0